refactor: satisfy runtime integration guardrails
